requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
torch
sentence-transformers
huggingface-hub

python-dotenv

numpy
scikit-learn

streamlit